+++ daemon/list.go
@@ -7,8 +7,8 @@ import (
 	"strings"
 
 	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/nat"
 	"github.com/docker/docker/pkg/graphdb"
+	"github.com/docker/docker/pkg/nat"
 	"github.com/docker/docker/pkg/parsers/filters"
 )
