+++ runconfig/config.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"strings"
 
-	"github.com/docker/docker/nat"
+	"github.com/docker/docker/pkg/nat"
 )
 
 // Entrypoint encapsulates the container entrypoint.
