+++ runconfig/merge.go
@@ -3,7 +3,7 @@ package runconfig
 import (
 	"strings"
 
-	"github.com/docker/docker/nat"
+	"github.com/docker/docker/pkg/nat"
 )
 
 func Merge(userConf, imageConf *Config) error {
